get rid of stopOnLimit
Because of stepBack (removing a previously placed shift) stopOnLimit was
causing a never-ending loop for some data sets where the was a ping-
ponging of adding/removing shifts. This is because when relying on the
number of partitions rather than the improvement in model fit, worse
models are added, but then removed (because they are bad), but then
added again (because something *has* to be added), etc.

In doing this, 'partitions' has taken a new meaning. Rather than
specifying *the* number of piecewise models to fit, it is now a maximum.
This may be used when, say, someone wants to get just the first 10 models
fit to their tree. However, this is just a suggested maximum: if model
improvement stops, or the MEDUSA-determined maximum number of models has
been encountered (i.e. when denominator of aicc correction is undefined),
then the algorithm stops.

So, stopOnThreshold/stopOnLimit flavours have been combined into a single
function. Easier to read, does what people will want.